Handle Last Logon Correctly
Previously we did not allow for the last logon time being measured to the millisecond.
This update now allows for that.

[resolves #822]

+++ cloudfoundry-client/src/main/java/org/cloudfoundry/uaa/users/_UserInfoResponse.java
@@ -64,7 +64,7 @@ abstract class _UserInfoResponse {
      */
     @JsonProperty("previous_logon_time")
     @Nullable
-    abstract Integer getPreviousLogonTime();
+    abstract Long getPreviousLogonTime();
 
     /**
      * The user sub
